ý¿¿¹°Added updateLS function.
This function relies on global state to check if the session has timed out via
the SESSION_TIMEOUT variable. The `checkTimer()` function has also been
refactored to use this global state rather than a function call.

Additionally, a worker is set to check the DOM for a div indicating timeout
every `TIMEOUT_INTERVAL/6e4` minutes.

All of this is cleaned up on tab closure and a mechanism is in place for newer
tabs to resume this operations once the master tab has been closed.